Parte lógica melhorada
Desenvolvam os JUnit, lembre-se das regras:
1- Testes independentes
2-Nomes funcionais aos testes
3-Testem o necessário

+++ Codigo/persistence/ProfessorDAO.java
@@ -9,8 +9,13 @@ import exception.ClienteException;
 
 public class ProfessorDAO {
 
-	private static final String PROFESSOR_JA_EXISTENTE = "O Professor ja esta cadastrado.";
-	private static final String PROFESSOR_NAO_EXISTENTE = "O Professor nao esta cadastrado.";
+	//Mensagens
+		private static final String PROFESSOR_JA_EXISTENTE = "O Professor ja esta cadastrado.";
+		private static final String PROFESSOR_NAO_EXISTENTE = "O Professor nao esta cadastrado.";
+		private static final String PROFESSOR_NULO = "O Aluno esta nulo.";
+		private static final String PROFESSOR_EM_USO = "Sala esta sendo utilizada em uma reserva.";
+		private static final String CPF_JA_EXISTENTE = "Ja existe um aluno cadastrado com esse CPF.";
+		private static final String MATRICULA_JA_EXISTENTE = "Ja existe um aluno cadastrado com essa matricula.";
 	
 	//Singleton
 		private static ProfessorDAO instance;
@@ -23,9 +28,15 @@ public class ProfessorDAO {
 		}
 	//
 	
+	
 	public void incluir(Professor prof) throws SQLException, ClienteException {
-			
-			if(!this.inDB(prof))
+			if(prof == null)
+				throw new ClienteException(PROFESSOR_NULO);
+			else if(this.inDBCpf(prof.getCpf()))
+				throw new ClienteException(CPF_JA_EXISTENTE);
+			else if(this.inDBMatricula(prof.getMatricula()))
+				throw new ClienteException(MATRICULA_JA_EXISTENTE);
+			else if(!this.inDB(prof))
 			{
 				this.updateQuery("INSERT INTO " +
 						"professor (nome, cpf, telefone, email, matricula) VALUES (" +
@@ -36,19 +47,27 @@ public class ProfessorDAO {
 						"\"" + prof.getMatricula() + "\"); "
 					);
 			}
-			else {
+			else
 				throw new ClienteException(PROFESSOR_JA_EXISTENTE);
-			}
 			
 	}
 
 	public void alterar(Professor prof_velho, Professor prof_novo) throws SQLException, ClienteException {
+		if(prof_velho == null)
+			throw new ClienteException(PROFESSOR_NULO);
+		if(prof_novo == null)
+			throw new ClienteException(PROFESSOR_NULO);
 		Connection con = FactoryConnection.getInstance().getConnection();
 		PreparedStatement pst;
 		
-		if(!this.inDB(prof_velho)){
+		if(!this.inDB(prof_velho))
 			throw new ClienteException(PROFESSOR_NAO_EXISTENTE);
-		}
+		if(this.inOtherDB(prof_velho))
+			throw new ClienteException(PROFESSOR_EM_USO);
+		else if(this.inDBCpf(prof_novo.getCpf()))
+			throw new ClienteException(CPF_JA_EXISTENTE);
+		else if(this.inDBMatricula(prof_novo.getMatricula()))
+			throw new ClienteException(MATRICULA_JA_EXISTENTE);
 		else if(!this.inDB(prof_novo)){
 			String msg = "UPDATE professor SET " +
 					"nome = \"" + prof_novo.getNome() + "\", " +
@@ -66,16 +85,19 @@ public class ProfessorDAO {
 			pst = con.prepareStatement(msg);
 			pst.executeUpdate();
 			con.commit();
-		}else {
+		}else
 			throw new ClienteException(PROFESSOR_JA_EXISTENTE);
-		}
 		
 		pst.close();
 		con.close();
 	}
 
 	public void excluir(Professor prof) throws SQLException, ClienteException {
-		if(this.inDB(prof)){
+		if(prof == null)
+			throw new ClienteException(PROFESSOR_NULO);
+		if(this.inOtherDB(prof))
+			throw new ClienteException(PROFESSOR_EM_USO);
+		else if(this.inDB(prof)){
 			this.updateQuery("DELETE FROM professor WHERE " +
 				"professor.nome = \"" + prof.getNome() + "\" and " +
 				"professor.cpf = \"" + prof.getCpf() + "\" and " +
@@ -88,12 +110,38 @@ public class ProfessorDAO {
 			throw new ClienteException(PROFESSOR_NAO_EXISTENTE);
 	}
 
+	
+	
 	public Vector<Professor> buscarTodos() throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor;");
+	}
+	public Vector<Professor> buscarNome(String valor) throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor WHERE nome = " + "\"" + valor + "\";");
+	}
+	public Vector<Professor> buscarCpf(String valor) throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor WHERE cpf = " + "\"" + valor + "\";");
+	}
+	public Vector<Professor> buscarMatricula(String valor) throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor WHERE matricula = " + "\"" + valor + "\";");
+	}
+	public Vector<Professor> buscarEmail(String valor) throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor WHERE email = " + "\"" + valor + "\";");
+	}
+	public Vector<Professor> buscarTelefone(String valor) throws SQLException, ClienteException {
+		return this.buscar("SELECT * FROM professor WHERE telefone = " + "\"" + valor + "\";");
+	}
+
+	
+	/**
+	 * Metodos Privados
+	 * */
+	
+	private Vector<Professor> buscar(String query) throws SQLException, ClienteException {
 		Vector<Professor> vet = new Vector<Professor>();
 		
 		Connection con =  FactoryConnection.getInstance().getConnection();
 		
-		PreparedStatement pst = con.prepareStatement("SELECT * FROM professor;");
+		PreparedStatement pst = con.prepareStatement(query);
 		ResultSet rs = pst.executeQuery();
 		
 		while(rs.next())
@@ -105,7 +153,8 @@ public class ProfessorDAO {
 		return vet;
 	}
 	
-	public boolean inDB(Professor prof) throws SQLException{
+	
+	private boolean inDB(Professor prof) throws SQLException{
 		Connection con = FactoryConnection.getInstance().getConnection();
 		PreparedStatement pst = con.prepareStatement("SELECT * FROM professor WHERE " +
 				"professor.nome = \"" + prof.getNome() + "\" and " +
@@ -128,11 +177,91 @@ public class ProfessorDAO {
 			return true;
 		}
 	}
-	
-	public Professor buscar(Professor prof) throws SQLException {
-		//TODO
-		return null;
+	private boolean inDBCpf(String codigo) throws SQLException{
+		Connection con = FactoryConnection.getInstance().getConnection();
+		PreparedStatement pst = con.prepareStatement("SELECT * FROM professor WHERE " +
+				"cpf = \"" + codigo + "\";");
+		ResultSet rs = pst.executeQuery();
+		
+		if(!rs.next())
+		{
+			rs.close();
+			pst.close();
+			con.close();
+			return false;
+		}
+		else {
+			rs.close();
+			pst.close();
+			con.close();
+			return true;
+		}
 	}
+	private boolean inDBMatricula(String codigo) throws SQLException{
+		Connection con = FactoryConnection.getInstance().getConnection();
+		PreparedStatement pst = con.prepareStatement("SELECT * FROM professor WHERE " +
+				"matricula = \"" + codigo + "\";");
+		ResultSet rs = pst.executeQuery();
+		
+		if(!rs.next())
+		{
+			rs.close();
+			pst.close();
+			con.close();
+			return false;
+		}
+		else {
+			rs.close();
+			pst.close();
+			con.close();
+			return true;
+		}
+	}
+	private boolean inOtherDB(Professor prof) throws SQLException{
+		Connection con =  FactoryConnection.getInstance().getConnection();
+		
+		PreparedStatement pst = con.prepareStatement(
+				"SELECT * FROM reserva_sala_professor WHERE " +
+				"id_professor = (SELECT id_professor FROM professor WHERE " +
+				"professor.nome = \"" + prof.getNome() + "\" and " +
+				"professor.cpf = \"" + prof.getCpf() + "\" and " +
+				"professor.telefone = \"" + prof.getTelefone() + "\" and " +
+				"professor.email = \"" + prof.getEmail() + "\" and " +
+				"professor.matricula = \"" + prof.getMatricula() + "\");");
+		ResultSet rs = pst.executeQuery();
+		
+		if(rs.next())
+		{
+			rs.close();
+			pst.close();
+			con.close();
+			return true;
+		}
+		
+		pst = con.prepareStatement(
+				"SELECT * FROM reserva_equipamento WHERE " +
+				"id_professor = (SELECT id_professor FROM professor WHERE " +
+				"professor.nome = \"" + prof.getNome() + "\" and " +
+				"professor.cpf = \"" + prof.getCpf() + "\" and " +
+				"professor.telefone = \"" + prof.getTelefone() + "\" and " +
+				"professor.email = \"" + prof.getEmail() + "\" and " +
+				"professor.matricula = \"" + prof.getMatricula() + "\");");
+		rs = pst.executeQuery();
+		
+		if(rs.next())
+		{
+			rs.close();
+			pst.close();
+			con.close();
+			return true;
+		}
+		
+		rs.close();
+		pst.close();
+		con.close();
+		return false;
+	}
+	
 	
 	private Professor fetchProfessor(ResultSet rs) throws ClienteException, SQLException{
 		return new Professor(rs.getString("nome"), rs.getString("cpf"), rs.getString("matricula"),
